+++ crates/client/tests/expected_source_codes/expected_test_fuzz.rs
@@ -1,25 +1,26 @@
 use trident_client::fuzzing::*;
 mod fuzz_instructions;
 use dummy_2::entry as entry_dummy_2;
-use dummy_2::ID as PROGRAM_ID_DUMMY_2;
+use dummy_2::ID as PROGRAM_ID_dummy_2;
 use dummy_example::entry as entry_dummy_example;
-use dummy_example::ID as PROGRAM_ID_DUMMY_EXAMPLE;
+use dummy_example::ID as PROGRAM_ID_dummy_example;
 use fuzz_instructions::FuzzInstruction;
-const PROGRAM_NAME_DUMMY_2: &str = "dummy_2";
-const PROGRAM_NAME_DUMMY_EXAMPLE: &str = "dummy_example";
-struct MyFuzzData;
+const PROGRAM_NAME_dummy_2: &str = "dummy_2";
+const PROGRAM_NAME_dummy_example: &str = "dummy_example";
+struct InstructionsSequence;
 /// Define instruction sequences for invocation.
-/// `pre_ixs` runs at the start, `ixs` in the middle, and `post_ixs` at the end.
-/// For example, to call `InitializeFn` at the start of each fuzzing iteration:
+/// `pre` runs at the start, `middle` in the middle, and `post` at the end.
+/// For example, to call `InitializeFn`, `UpdateFn` and then `WithdrawFn` during
+/// each fuzzing iteration:
 /// ```
-/// fn pre_ixs(u: &mut arbitrary::Unstructured) ->
-/// arbitrary::Result<Vec<FuzzInstruction>> {
-///     let init = FuzzInstruction::InitializeFn(InitializeFn::arbitrary(u)?);
-///     Ok(vec![init])
-/// }
+/// use fuzz_instructions::{InitializeFn, UpdateFn, WithdrawFn};
+/// impl FuzzDataBuilder<FuzzInstruction> for InstructionsSequence {
+///     pre_sequence!(InitializeFn,UpdateFn);
+///     middle_sequence!(WithdrawFn);
+///}
 /// ```
-/// For more details, see: https://ackee.xyz/trident/docs/dev/features/instructions-sequences/#instructions-sequences
-impl FuzzDataBuilder<FuzzInstruction> for MyFuzzData {}
+/// For more details, see: https://ackee.xyz/trident/docs/latest/features/instructions-sequences/#instructions-sequences
+impl FuzzDataBuilder<FuzzInstruction> for InstructionsSequence {}
 /// `fn fuzz_iteration` runs during every fuzzing iteration.
 /// Modification is not required.
 fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
@@ -27,13 +28,13 @@ fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
     config: &Config,
 ) {
     let fuzzing_program_dummy_2 = FuzzingProgram::new(
-        PROGRAM_NAME_DUMMY_2,
-        &PROGRAM_ID_DUMMY_2,
+        PROGRAM_NAME_dummy_2,
+        &PROGRAM_ID_dummy_2,
         processor!(convert_entry!(entry_dummy_2)),
     );
     let fuzzing_program_dummy_example = FuzzingProgram::new(
-        PROGRAM_NAME_DUMMY_EXAMPLE,
-        &PROGRAM_ID_DUMMY_EXAMPLE,
+        PROGRAM_NAME_dummy_example,
+        &PROGRAM_ID_dummy_example,
         processor!(convert_entry!(entry_dummy_example)),
     );
     let mut client = ProgramTestClientBlocking::new(
@@ -45,5 +46,5 @@ fn fuzz_iteration<T: FuzzTestExecutor<U> + std::fmt::Display, U>(
 }
 fn main() {
     let config = Config::new();
-    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : MyFuzzData | { fuzz_iteration (fuzz_data , & config) ; });
+    fuzz_trident ! (fuzz_ix : FuzzInstruction , | fuzz_data : InstructionsSequence | { fuzz_iteration (fuzz_data , & config) ; });
 }