+++ crates/fuzz/src/program_test_client_blocking.rs
@@ -3,14 +3,13 @@ use solana_program_test::ProgramTest;
 use solana_program_test::ProgramTestContext;
 use solana_sdk::account::Account;
 use solana_sdk::account_info::AccountInfo;
+use solana_sdk::clock::Clock;
 use solana_sdk::entrypoint::ProgramResult;
-use solana_sdk::system_program::ID as SYSTEM_PROGRAM_ID;
+use solana_sdk::sysvar::Sysvar;
 use solana_sdk::{
-    account::AccountSharedData, hash::Hash, instruction::AccountMeta, program_option::COption,
-    program_pack::Pack, pubkey::Pubkey, rent::Rent, signature::Keypair, signature::Signer,
+    account::AccountSharedData, hash::Hash, pubkey::Pubkey, rent::Rent, signature::Keypair,
     transaction::VersionedTransaction,
 };
-use spl_token::state::Mint;
 use tokio::runtime::Builder;
 
 use crate::config::Config;
@@ -105,125 +104,27 @@ macro_rules! convert_entry {
 }
 
 impl FuzzClient for ProgramTestClientBlocking {
-    fn set_account(&mut self, lamports: u64) -> Keypair {
-        let owner = Keypair::new();
-        let account = AccountSharedData::new(lamports, 0, &SYSTEM_PROGRAM_ID);
-        self.ctx.set_account(&owner.pubkey(), &account);
-        owner
-    }
-
-    fn set_token_account(
-        &mut self,
-        mint: Pubkey,
-        owner: Pubkey,
-        amount: u64,
-        delegate: Option<Pubkey>,
-        is_native: Option<u64>,
-        delegated_amount: u64,
-        close_authority: Option<Pubkey>,
-    ) -> Pubkey {
-        let mint_account_key = Keypair::new().pubkey();
-
-        let delegate = match delegate {
-            Some(a) => COption::Some(a),
-            _ => COption::None,
-        };
-
-        let is_native = match is_native {
-            Some(a) => COption::Some(a),
-            _ => COption::None,
-        };
-
-        let close_authority = match close_authority {
-            Some(a) => COption::Some(a),
-            _ => COption::None,
-        };
-
-        let r = Rent::default();
-        let lamports = r.minimum_balance(spl_token::state::Account::LEN);
-
-        let mut account =
-            AccountSharedData::new(lamports, spl_token::state::Account::LEN, &spl_token::id());
-
-        let token_account = spl_token::state::Account {
-            mint,
-            owner,
-            amount,
-            delegate,
-            state: spl_token::state::AccountState::Initialized,
-            is_native,
-            delegated_amount,
-            close_authority,
-        };
-
-        let mut data = vec![0u8; spl_token::state::Account::LEN];
-        spl_token::state::Account::pack(token_account, &mut data[..]).unwrap();
-        account.set_data_from_slice(&data);
-        self.ctx.set_account(&mint_account_key, &account);
-
-        mint_account_key
-    }
-
-    fn set_mint_account(
-        &mut self,
-        decimals: u8,
-        owner: &Pubkey,
-        freeze_authority: Option<Pubkey>,
-    ) -> Pubkey {
-        let mint_account = Keypair::new();
-
-        let authority = match freeze_authority {
-            Some(a) => COption::Some(a),
-            _ => COption::None,
-        };
-
-        let r = Rent::default();
-        let lamports = r.minimum_balance(Mint::LEN);
-
-        let mut account = AccountSharedData::new(lamports, Mint::LEN, &spl_token::id());
-
-        let mint = Mint {
-            is_initialized: true,
-            mint_authority: COption::Some(*owner),
-            freeze_authority: authority,
-            decimals,
-            ..Default::default()
-        };
-
-        let mut data = vec![0u8; Mint::LEN];
-        Mint::pack(mint, &mut data[..]).unwrap();
-        account.set_data_from_slice(&data);
-        self.ctx.set_account(&mint_account.pubkey(), &account);
-
-        mint_account.pubkey()
-    }
-
     fn payer(&self) -> Keypair {
         self.ctx.payer.insecure_clone()
     }
 
-    fn get_account(&mut self, key: &Pubkey) -> Result<Option<Account>, FuzzClientError> {
-        Ok(self
+    fn get_account(&mut self, key: &Pubkey) -> AccountSharedData {
+        let account = self
             .rt
             .block_on(self.ctx.banks_client.get_account_with_commitment(
                 *key,
                 solana_sdk::commitment_config::CommitmentLevel::Confirmed,
-            ))?)
+            ))
+            .unwrap_or_default();
+        match account {
+            Some(account) => account.into(),
+            None => {
+                let account = AccountSharedData::new(0, 0, &solana_sdk::system_program::ID);
+                self.ctx.set_account(key, &account);
+                account
+            }
+        }
     }
-    fn get_accounts(
-        &mut self,
-        metas: &[AccountMeta],
-    ) -> Result<Vec<Option<Account>>, FuzzClientErrorWithOrigin> {
-        let result: Vec<_> = metas
-            .iter()
-            .map(|m| {
-                self.get_account(&m.pubkey)
-                    .map_err(|e| e.with_origin(Origin::Account(m.pubkey)))
-            })
-            .collect();
-        result.into_iter().collect()
-    }
-
     fn get_last_blockhash(&self) -> Hash {
         self.ctx.last_blockhash
     }
@@ -240,7 +141,31 @@ impl FuzzClient for ProgramTestClientBlocking {
         self.ctx.set_account(address, account);
     }
 
-    fn get_rent(&mut self) -> Result<Rent, FuzzClientError> {
-        Ok(self.rt.block_on(self.ctx.banks_client.get_rent())?)
+    fn forward_in_time(&mut self, seconds: i64) -> Result<(), FuzzClientError> {
+        // Get the current clock state from the program test context.
+        let mut clock = self
+            .rt
+            .block_on(self.ctx.banks_client.get_sysvar::<Clock>())?;
+
+        // Calculate the new timestamp after advancing time.
+        let new_timestamp = clock.unix_timestamp.saturating_add(seconds);
+
+        // Update the Clock instance with the new timestamp.
+        clock.unix_timestamp = new_timestamp;
+
+        // Update the sysvar in the program test context with the new Clock state.
+        self.ctx.set_sysvar(&clock);
+        Ok(())
+    }
+    fn warp_to_slot(&mut self, warp_slot: u64) {
+        let _ = self.ctx.warp_to_slot(warp_slot);
+    }
+    fn warp_to_epoch(&mut self, warp_epoch: u64) {
+        let _ = self.ctx.warp_to_epoch(warp_epoch);
+    }
+    fn get_sysvar<T: Sysvar>(&mut self) -> T {
+        self.rt
+            .block_on(self.ctx.banks_client.get_sysvar::<T>())
+            .unwrap_or_default()
     }
 }