+++ examples/fuzz-tests/unchecked-arithmetic-0/programs/unchecked-arithmetic-0/src/lib.rs
@@ -2,8 +2,6 @@ use anchor_lang::prelude::*;
 
 const MAGIC_NUMBER: u8 = 254;
 
-use trident_derive_accounts_snapshots::AccountsSnapshots;
-
 declare_id!("BM8vocQeC2VuDf1KhbHLsZxTh7owzDNTAkKyZoTxFiUs");
 
 #[program]
@@ -38,7 +36,7 @@ pub fn buggy_math_function(input1: u8, input2: u8) -> u8 {
     input1 / divisor
 }
 
-#[derive(Accounts, AccountsSnapshots)]
+#[derive(Accounts)]
 pub struct Initialize<'info> {
     #[account(init, payer = user, space = 8 + 40)]
     pub counter: Account<'info, Counter>,
@@ -49,7 +47,7 @@ pub struct Initialize<'info> {
     pub system_program: Program<'info, System>,
 }
 
-#[derive(Accounts, AccountsSnapshots)]
+#[derive(Accounts)]
 pub struct Update<'info> {
     #[account(mut, has_one = authority)]
     pub counter: Account<'info, Counter>,